利用JpaRepository实现的mvc模型

+++ 21springmvcdemo/src/main/java/springmvcdemo/authentication/model/User.java
@@ -7,18 +7,17 @@ import javax.persistence.Table;
 @Entity
 @Table(name="users")
 public class User {
-	
-	private int id;
+	private Long id;
     private String username;
     private String password;
     private String role;
     private Boolean enabled;
     
     @Id
-    public int getId() {
+    public Long getId() {
         return id;
     }
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
